Refactor BitmapExt into BitmapBase and BitmapExt. Extract avx_support out of avx.
