+++ DeepSkyStacker/DeepSkyStacker.vcxproj.filters
@@ -369,6 +369,9 @@
     <ClCompile Include="avx_entropy.cpp">
       <Filter>Kernel</Filter>
     </ClCompile>
+    <ClCompile Include="avx_support.cpp">
+      <Filter>Kernel</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="DeepSkyStacker.rc">
@@ -743,6 +746,15 @@
       <Filter>Kernel</Filter>
     </ClInclude>
     <ClInclude Include="avx_entropy.h">
+      <Filter>Kernel</Filter>
+    </ClInclude>
+    <ClInclude Include="BitmapBase.h">
+      <Filter>Kernel</Filter>
+    </ClInclude>
+    <ClInclude Include="RefCount.h">
+      <Filter>Tools</Filter>
+    </ClInclude>
+    <ClInclude Include="avx_support.h">
       <Filter>Kernel</Filter>
     </ClInclude>
   </ItemGroup>
