+++ DeepSkyStacker/avx_luminance.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "BitmapExt.h"
+#include "BitmapBase.h"
 #include <tuple>
 
 class AvxLuminance
