+++ DeepSkyStacker/avx_output.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "BitmapExt.h"
+#include "BitmapBase.h"
 #include <vector>
