+++ Tools/SmartPtr.h
@@ -1,33 +1,5 @@
 #ifndef	_SMARTPTR_H_
 #define	_SMARTPTR_H_
-#include <atomic>
-
-class CRefCount
-{
-protected:
-	std::atomic<long> m_lRefCount;
-
-public:
-	CRefCount() :
-		m_lRefCount{ 0 }
-	{
-	};
-
-	virtual ~CRefCount()
-	{
-	};
-
-	void	AddRef()
-	{
-		m_lRefCount++;
-	};
-
-	void	Release()
-	{
-		if (m_lRefCount.fetch_sub(1) == 1) // If it was previously one (is now zero), delete the object.
-			delete this;
-	};
-};
 
 template <class T>
 class CSmartPtr
